+++ CHANGES.markdown
@@ -1,4 +1,5 @@
-## Next
+## Changes in 1.5.0
+ - Add the `example` function, it fixes the type of an Expectation (#86)
  - Rename `--fast-fail` to `--fail-fast` (for consistency with RSpec)
  - Do not suppress output to `stdout`